requirements.txt
picamera2
flask
flask-cors
azure-storage-blob
face_recognition
opencv-python-headless
firebase-admin
requests
waitress
psutil
